REFACTOR: toolgen files code splitting
