+++ lib/screens/common_widgets/agentic_ui_features/ai_ui_designer/generate_ui_dialog.dart
@@ -75,6 +75,7 @@ class _GenerateUIDialogState extends ConsumerState<GenerateUIDialog> {
         backgroundColor: Colors.redAccent,
       ));
       Navigator.pop(context);
+      return null;
     }
   }
