+++ packages/genai/test/genai_test.dart
@@ -1,5 +1,1 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:genai/genai.dart';
-
-void main() {
-}
+void main() {}
